feat(viewmodel): switch to local Room, add on-device search and face clustering

+++ android/app/src/main/java/com/pixelvault/app/ui/navigation/Screen.kt
@@ -11,6 +11,6 @@ sealed class Screen(val route: String) {
     }
 
     data object PersonPhotos : Screen("person/{clusterId}") {
-        fun create(clusterId: Int) = "person/$clusterId"
+        fun create(clusterId: Long) = "person/$clusterId"
     }
 }
